+++ flutter/android/src/main/kotlin/io/sentry/flutter/SentryFlutterPlugin.kt
@@ -15,6 +15,7 @@ import io.sentry.HubAdapter
 import io.sentry.SentryEvent
 import io.sentry.SentryLevel
 import io.sentry.Sentry
+import io.sentry.DateUtils
 import io.sentry.android.core.ActivityFramesTracker
 import io.sentry.android.core.AppStartState
 import io.sentry.android.core.LoadClass
@@ -168,7 +169,7 @@ class SentryFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
         val version = sdk["version"] as? String
         if (name != null && version != null) {
           val sdkVersion = SdkVersion(name, version)
-          options.setSentryClientName(name)
+          options.setSentryClientName("$name/$version")
           options.setSdkVersion(sdkVersion)
         }
       }
@@ -199,8 +200,9 @@ class SentryFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
       Log.w("Sentry", "App start won't be sent due to missing isColdStart")
       result.success(null)
     } else {
+      val appStartTimeMillis = DateUtils.nanosToMillis(appStartTime.nanoTimestamp().toDouble())
       val item = mapOf<String, Any?>(
-        "appStartTime" to appStartTime.time.toDouble(),
+        "appStartTime" to appStartTimeMillis,
         "isColdStart" to isColdStart
       )
       result.success(item)